Add modularity to the js logic(extract functions responsible for verdict and modal to separate files)
