+++ index.html
@@ -11,7 +11,6 @@
     <link rel="stylesheet" href="index.css" />
 
     <link rel="icon" type="image/x-icon" href="./images/favicon.ico" />
-    <script src="js/app.js" defer></script>
     <title>Tic Tac Toe</title>
   </head>
   <body>
@@ -97,5 +96,7 @@
         <button class="restart-cta-btn" data-modal="restart"></button>
       </div>
     </div>
+
+    <script type="module" src="./js/app.js" defer></script>
   </body>
 </html>
